requirements.txt
google-genai
python-dotenv>=1.0.0
pydantic>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pytector>=0.1.0

# Optional dependencies (uncomment if needed)
#streamlit>=1.28.0
